+++ CHANGELOG.md
@@ -15,6 +15,8 @@ keybinds if they contained either `{` or `}`. You will now need to escape these 
 
 ### Fixed
 
+- Theme hot reload not working when it was set using the `theme.ron` form
+
 ## [0.11.0] - 2026-02-01
 
 ### Added